Refactored multistep components
