+++ packages/views/screens/managers/hwWalletLogin/hwWalletLogin.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
 import { subscribeToDevicesList } from '@wallet/utilities/hwManager';
 import routes from '@screens/router/routes';
-import MultiStep from '@shared/multiStep';
+import MultiStep from 'src/modules/common/components/OldMultiStep';
 import Loading from './loading';
 import RequestPin from './requestPin';
 import SelectAccount from './selectAccount';
