+++ packages/views/screens/managers/reclaimBalance/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { withTranslation } from 'react-i18next';
-import MultiStep from '@shared/multiStep';
+import MultiStep from 'src/modules/common/components/OldMultiStep';
 import Introduction from './introduction';
 import Reclaim from './reclaim';
 import styles from './index.css';
