+++ packages/views/screens/managers/signMultiSignTransaction/index.js
@@ -1,7 +1,7 @@
 /* istanbul ignore file */
 import React from 'react';
 import { withRouter } from 'react-router-dom';
-import MultiStep from '@shared/multiStep';
+import MultiStep from 'src/modules/common/components/OldMultiStep';
 import { removeSearchParamsFromUrl } from '@screens/router/searchParams';
 import TransactionSignature from '@transaction/detail/manager/transactionSignature';
 import Dialog from '@basics/dialog/dialog';
