+++ src/modules/common/components/MultiStep/Button.js
@@ -1,13 +1,6 @@
 import React from 'react';
 
-export const Element = ({ children, type, ...rest }) => {
-  if (typeof document !== 'undefined') {
-    return <div {...rest}>{children}</div>;
-  }
-  return null;
-};
-
-export const Button = ({
+const Button = ({
   children, type, onClick, ...rest
 }) => {
   if (typeof document !== 'undefined') {
@@ -19,3 +12,5 @@ export const Button = ({
   }
   return null;
 };
+
+export default Button;
